+++ src/main/java/foodfinder/dto/History.java
@@ -1,7 +1,6 @@
 package foodfinder.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -10,7 +9,6 @@ import javax.validation.constraints.NotNull;
 import java.sql.Timestamp;
 
 @Entity
-@Builder
 @AllArgsConstructor
 @NotNull
 @NoArgsConstructor
