+++ src/main/java/foodfinder/dto/RestaurantGrade.java
@@ -1,7 +1,6 @@
 package foodfinder.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -9,7 +8,6 @@ import org.springframework.stereotype.Service;
 import javax.persistence.*;
 
 @Entity
-@Builder
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -21,7 +19,6 @@ public class RestaurantGrade {
     @SequenceGenerator(name = "grade_seq", sequenceName = "seq_grade", initialValue = 1, allocationSize = 1)
     @Id
     @Column(name = "id_grade")
-
     private Integer idGrade;
     private Integer grade;
     @Column(name = "id_restaurants")
