+++ src/main/java/foodfinder/dto/StatusResponse.java
@@ -1,12 +1,10 @@
 package foodfinder.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.http.HttpStatus;
 
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
