+++ src/main/java/foodfinder/dto/User.java
@@ -8,8 +8,6 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 
-
-@Builder
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -27,8 +25,4 @@ public class User {
     private String surname;
     private String mail;
     private String password;
-
-
-
-
 }
